requirements.txt
streamlit>=1.45.0
python-dotenv>=1.0.0
openai>=1.2.0
pydantic>=2.7
backoff>=2.2
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.10.0 